Created the Card Object
Created the initial Card object that will store the details of a single
card in the Deck.

Attributes:

suitName - Name of the card's suit
suitValue - Value assigned to the suit
rankName - Name of the card's rank
rankValue - Value assigned to the rank

For the values I have designed to assign ascending prime numbers to each
rank, the higher the prime number - the higher the effective value of
the card. These prime numbers will aid in the various checks needed to
determine a winning hand