Created the Hand Object
This is required for a variety of situations.

Players will have hole cards (hand of two cards). Players will also
share the community cards (5 table cards).

The hand object will also be used to serve as a combination of the
above, as the aim of poker is to make the best 5 card hand from your
hole cards and community cards.